test(core): migrate UpdateGhosttyTests (Wave C)
24 tests, 74 failure sites preserved 1:1. Straight #expect conversion -- no `throw TestFailure` sites in the legacy file.
